+++ express/blocks/carousel-card-mobile/carousel-card-mobile.js
@@ -10,12 +10,8 @@
  * governing permissions and limitations under the License.
  */
 
-import {
-  buildStaticFreePlanWidget,
-  createTag,
-  fetchRelevantRows,
-} from '../../scripts/scripts.js';
-
+import { createTag, fetchRelevantRows } from '../../scripts/scripts.js';
+import { buildStaticFreePlanWidget } from '../../scripts/utils/free-plan.js';
 import buildPaginatedCarousel from '../shared/paginated-carousel.js';
 import { buildAppStoreBadge } from '../shared/app-store-badge.js';
