+++ express/blocks/fullscreen-marquee-desktop/fullscreen-marquee-desktop.js
@@ -10,12 +10,12 @@
  * governing permissions and limitations under the License.
  */
 import {
-  addFreePlanWidget,
   createOptimizedPicture,
   createTag,
   fetchPlaceholders,
   transformLinkToAnimation,
 } from '../../scripts/scripts.js';
+import { addFreePlanWidget } from '../../scripts/utils/free-plan.js';
 
 function buildContent(content) {
   const contentLink = content.querySelector('a');
