+++ express/blocks/marquee/marquee.js
@@ -17,9 +17,9 @@ import {
   getLocale,
   addHeaderSizing,
   getIconElement,
-  addFreePlanWidget,
   fetchPlaceholders,
 } from '../../scripts/scripts.js';
+import { addFreePlanWidget } from '../../scripts/utils/free-plan.js';
 
 const breakpointConfig = [
   {
